week 4/RAG/requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
faiss-cpu
python-dotenv
pydantic
aiofiles
requests
nltk
